requirements.txt
Flask==3.0.0
flask-restx==1.3.0
flask-limiter==3.5.0
mongoengine==0.27.0
pymongo==4.6.0
firebase-admin==6.2.0
PyJWT==2.8.0
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
flask
flask-limiter
flask-restx
pymongo
gunicorn
flask-mail
flask-mail
Flask-Mail
firebase-admin
flask-mongoengine
pymongo
numpy
opencv-python
Pillow
PyMuPDF
PyPDF2
pytesseract
Pillow
PyMuPDF
